Update Moshkov Work
Moved old working files to "Seal Archive" folder, in case they want to be kept. Added new working notebook which corresponds with the work in the research report, as well as outputted results csv files from that notebook. Need to upload "assay_cp_data_PUMA.csv" somewhere externally as it can't be uploaded to GitHub due to size constraints (even if it's zipped it's over 100MB).